refactor(index): split entrypoint into app/server/services
Split src/index.ts (718 lines) into focused modules:

- app.ts: Express app factory (middleware, security, routes, probes, error handler)
- server.ts: HTTP server creation + WebSocket attachment
- services.ts: background-service startup orchestration

index.ts stays a thin entrypoint: tracer init, wiring, listen, and graceful
shutdown (incl. connection draining). Rebased onto current main so upstream
features — cookie auth, CSRF, compression, versioning, /healthz, key
rotation, and prismaBackfill — are preserved.

Fixes the Lint CI failure (8 unused-import errors caused by the previous
mismatched base).
